requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
slowapi>=0.1.7
azure-storage-blob[aio]>=12.18.0
redis>=5.0.0
aiohttp>=3.8.0
uvloop>=0.17.0